Adds support for not scheduled appointments. Fixes to constraints formatting in modal + event colors
Not scheduled appointments are caught in the Solver and sent as json to the client with full info about name/surname/house/requested constraints.
Client-side javascript parse this response and adds html elements at the bottom of the page to show not scheduled appointments.
Requested constraints in modal windows are displayed as bulleted lists as well.
//todo: Code to print solutions to json file with ordered keys available but commented in `SolverWithBacktrack.py` -- choose what to do.